requirements.txt
python-dotenv>=1.0.0
qdrant-client>=1.7.0
numpy>=1.24.0
modal>=0.65.0
tree-sitter>=0.21.0
tree-sitter-rust>=0.21.0
tree-sitter-typescript>=0.21.0
tree-sitter-solidity>=1.0.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0
openai>=1.0.0
